requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
slack_sdk==3.36.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
